Refactoring: using interface
