+++ app/src/main/java/com/dumdumbich/train/bubbles/domain/interactor/ICanvasView.kt
@@ -1,0 +1,7 @@
+package com.dumdumbich.train.bubbles.domain.interactor
+
+import com.dumdumbich.train.bubbles.domain.entity.MyCircle
+
+interface ICanvasView {
+    fun drawCircle(circle: MyCircle)
+}
