+++ app/src/main/java/com/dumdumbich/train/bubbles/ui/CanvasView.kt
@@ -1,28 +1,42 @@
-package com.dumdumbich.train.bubbles
+package com.dumdumbich.train.bubbles.ui
 
 import android.content.Context
 import android.graphics.Canvas
+import android.graphics.Paint
 import android.graphics.Point
 import android.util.AttributeSet
 import android.view.Display
 import android.view.View
 import android.view.WindowManager
+import com.dumdumbich.train.bubbles.domain.GameManager
+import com.dumdumbich.train.bubbles.domain.interactor.ICanvasView
+import com.dumdumbich.train.bubbles.domain.entity.MyCircle
 
-class CanvasView(context: Context, attrs: AttributeSet) : View(context, attrs) {
+
+class CanvasView(context: Context, attrs: AttributeSet) : View(context, attrs), ICanvasView {
 
     private var widthScreen = 0
     private var heightScreen = 0
-
+    private lateinit var paint: Paint
     private var gameManager: GameManager
+    private var canvas: Canvas? = null
 
     init {
+        initPaint()
         initScreenSize(context)
         gameManager = GameManager(this, widthScreen, heightScreen)
     }
 
+    private fun initPaint() {
+        paint = Paint()
+        paint.isAntiAlias = true
+        paint.style = Paint.Style.FILL
+    }
+
     override fun onDraw(canvas: Canvas?) {
         super.onDraw(canvas)
-        gameManager.onDraw(canvas)
+        this.canvas = canvas
+        gameManager.onDraw()
     }
 
     private fun initScreenSize(context: Context) {
@@ -33,6 +47,10 @@ class CanvasView(context: Context, attrs: AttributeSet) : View(context, attrs) {
         display.getSize(point)
         widthScreen = point.x
         heightScreen = point.y
+    }
+
+    override fun drawCircle(circle: MyCircle) {
+        canvas?.drawCircle(circle.x.toFloat(), circle.y.toFloat(), circle.r.toFloat(), paint)
     }
 
 }
