+++ app/src/main/java/com/dumdumbich/train/bubbles/ui/MainActivity.kt
@@ -1,7 +1,8 @@
-package com.dumdumbich.train.bubbles
+package com.dumdumbich.train.bubbles.ui
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import com.dumdumbich.train.bubbles.R
 
 class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
