+++ app/src/main/res/layout/activity_main.xml
@@ -4,9 +4,9 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#FFF"
-    tools:context=".MainActivity">
+    tools:context=".ui.MainActivity">
 
-    <com.dumdumbich.train.bubbles.CanvasView
+    <com.dumdumbich.train.bubbles.ui.CanvasView
         android:layout_width="match_parent"
         android:layout_height="match_parent" />
